pyproject.toml
[project]
name = "oulad"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "ipykernel>=7.2.0",
    "mostlyai-qa>=1.5.5",
    "numpy>=2.3.4",
    "pandas>=2.3.3",
    "scikit-learn>=1.7.2",
    "scipy>=1.15.3",
    "sdv>=1.29.1",
]
